+++ tests/test_template.nim
@@ -2,8 +2,5 @@ import unittest
 import crown_code
 
 suite "project bootstrap":
-  test "runApp proc compiles and runs":
-    crown_code.runApp()
-
   test "module can be imported":
     check true